+++ js/ckeck-max-length.js
@@ -1,0 +1,9 @@
+function checkMaxLength (string, maxString) {
+  if (typeof string === 'string' && string.length <= maxString) {
+    return true;
+  } else {
+    return false;
+  }
+}
+
+export { checkMaxLength };
